Convert usize to u63 in Store trait APIs
APIs in `StoreLike` now uses `u62` instead of `usize` It's because
`u64` expresses the size of the data, but `usize` expresses local
memory limits.

+++ nativelink-scheduler/src/memory_awaited_action_db.rs
@@ -83,7 +83,7 @@ impl Drop for ClientAwaitedAction {
 /// why the implementation has fixed default values in it.
 impl LenEntry for ClientAwaitedAction {
     #[inline]
-    fn len(&self) -> usize {
+    fn len(&self) -> u64 {
         0
     }
